Develop to master

+++ SmartStore.Application/Features/Products/Commands/UpdateProductCommand.cs
@@ -2,5 +2,5 @@
 
 namespace SmartStore.Application.Features.Products.Commands
 {
-    public record UpdateProductCommand(Guid Id, string Name, decimal Price) : IRequest<bool>;
+    public record UpdateProductCommand(Guid Id, string Name, double Price) : IRequest<bool>;
 }
